Creating paths for AppSecret

Create paths as a utility and set of rules/expectations. Useful when trying to enforce a convention.

app_secret_paths(appname = NULL, base_path = normalizePath("~"))

Arguments

appname
The application name, this will be prepended with a '.', and forced to be lowercase

base_path
The base path to use for storing the symmetric and encrypted files

Value

list()

Influential environment variables APP_SECRET_USE_HERE

See also

Examples

if (FALSE) {
  # list(symmetric_file = "/home/user/.best-shiny-app-ever/symmetric.rsa",
  #      key_file       = "/home/user/.best-shiny-app-ever/secret.pem")

  Sys.setenv("APP_SECRET_USE_HERE"="1")
  here::set_here(path = "/apps/another-shiny-app")
  # list(symmetric_file = "/apps/another-shiny-app/.user-settings/user/symmetric.rsa",
  #      key_file       = "/home/user/.best-shiny-app-ever/secret.pem")
  paths <- app_secret_paths(appname = "another-shiny-app")

  Sys.unsetenv("APP_SECRET_USE_HERE")
  here::set_here(path = "/apps/dull-app")
  # list(symmetric_file = "/apps/dull-app/.dull-app/user.symmetric.rsa",
  #      key_file       = "/home/user/.dull-app/secret.pem")
  paths <- app_secret_paths(appname = "dull-app", base_path = file.path(here()))
}