Application Secrets

A simple way to symmetrically encrypt information and store on disk.

Details

The goal of AppSecret is to manage application secrets for users, be they developers or test/production service users.

A PEM format public/private key pair is used to encrpyt a cryptographically random 32 bytes which is used as a 256bit key to encrypt any number of files, streams or data. The 256bit key and the encrpyted files may be stored with the application, either in the repository or on disk so long as the private key is kept secret. By default this is a user's home directory as defined by normalizePath("~").

Usage

paths <- app_secret_paths(appname = "client-app")

asm <- app_secret_manager(symmetric_file = paths$symmetric_file,
                          key_file       = paths$key_file)
## or
asm <- app_secret$new(symmetric_file = paths$symmetric_file,
                      key_file       = paths$key_file)

password_file <- asm$path_in_vault(filename = "client-password.enc")

decrypted_password <- asm$decrypt_file(file = password_file)

Arguments

appname A name for the manager to use in directory names

symmetric_file The path to the file that contains the symmetric 256bit key

key_file The path to the private key file in PEM format

filename The name of a file to store/access in the vault

file The absolute path to a file containing encrypted data

Methods

$new(symmetric_file, key_file) Initialise an instance of the app secrets manager with paths for the symmetric encryption password and private key file.

$create_keys(key_file) Create a public/private key pair in PEM format for later use. The key is immediately available for use by the manager.

decrypt_data(data) Decrypt data, which must be raw. The decrypted data will also be raw.

decrypt_file(file) Decrypt the contents of file returning a character vector.

encrypt_data(data) Encrypt data, which may be either of class character or raw. raw will be returned.

encrypt_file(file) Encrypt the contents of file.

path_in_vault(filename) Return an absolute path to a file with filename in the vault.

read_encrypted(file) Use readBin to slurp the contents of file.

set_debug(debug) Set the debug flag to the value of debug. Only useful for obtaining more information from errors at the current time.

tap(func) kestrel for injecting non-chaining methods func in a chain.

write_encrypted(data, file) Use writeBin to spurt data into file.

Author

Roy Storey roy.storey@plantandfood.co.nz

Examples

paths <- app_secret_paths(appname = "client-app")

asm <- app_secret_manager(symmetric_file = paths$symmetric_file,
                          key_file       = paths$key_file)

asm <- do.call(app_secret_manager, paths)

if (!file.exists(asm$path_in_vault("password.enc"))) {
  password <- "your password" ## or use getPass::getPass() /
                              ## rstudioapi::askForPassword()
  encrypted <- asm$encrypt_data(password)
  asm$write_encrypted(encrypted, asm$path_in_vault("password.enc"))
}

password <- asm$decrypt_file(asm$path_in_vault("password.enc"))

#> decrypting symmetric file success